requirements.txt
albumentations==1.2.1
matplotlib==3.7.3
numpy==1.22.4
pandas==1.4.3
pytorch_lightning==1.9.5
pyyaml==6.0
scikit_image==0.19.3
scikit_learn==1.3.2
scipy==1.8.1
torch==1.12.1+cu116
torchvision==0.13.1+cu116
torchaudio==0.12.1+cu116
opencv_contrib_python==4.4.0.46
seaborn==0.13.0
timm==0.6.12
torcheval==0.0.7
wandb==0.15.4
lightning_utilities==0.9.0
umap_learn==0.5.4
plotly
--extra-index-url https://download.pytorch.org/whl/cu116